ColorRamp class - Uses circular buffer to keep color history. - Uses SmoothVelocity3D to bounce within color space. Added SmootherVelocity3D Tweaked View Animated path and background based on color ramp

+++ src/com/anderson/Smoother.java
@@ -1,42 +1,42 @@
 package com.anderson;
 
 public class Smoother {
-    private final double deg_to_rad = 57.29576;
+    private final float deg_to_rad = 57.29576f;
 
-    private double radius;
-    private double degree;
-    private double speed;
+    private float radius;
+    private float degree;
+    private float speed;
     private boolean isSin;
-    private double direction;
+    private float direction;
 
-    public Smoother(boolean isSin, double radius, double speed) {
+    public Smoother(boolean isSin, float radius, float speed) {
         this.isSin = isSin;
         this.degree = 0;
         this.radius = radius;
         this.speed = speed;
-        this.direction = 1.0;
+        this.direction = 1.0f;
     }
 
     public void reverse() {
         this.direction = this.direction * -1;
     }
 
-    public void set_radius(double radius) {
+    public void set_radius(float radius) {
         this.radius = radius;
     }
 
     public void next_frame() {
         this.degree += this.speed;
         if (this.degree > 360.0) {
-            this.degree = this.degree % 360.0;
+            this.degree = this.degree % 360.0f;
         }
     }
 
-    public double get() {
+    public float get() {
         if (this.isSin) {
-            return Math.sin(this.degree / this.deg_to_rad) * this.radius * this.direction;
+            return (float)Math.sin(this.degree / this.deg_to_rad) * this.radius * this.direction;
         } else {
-            return Math.cos(this.degree / this.deg_to_rad) * this.radius * this.direction;
+            return (float)Math.cos(this.degree / this.deg_to_rad) * this.radius * this.direction;
         }
     }
